refactor: refactor command parsers to work with new implementation
