+++ src/main/java/seedu/address/logic/parser/AddDeadlineCommandParser.java
@@ -10,13 +10,14 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddDeadlineCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.entity.CalendarItemName;
-import seedu.address.model.entity.Deadline;
-import seedu.address.model.entity.Event;
-import seedu.address.model.entity.EventType;
-import seedu.address.model.entity.MatchableEvent;
-import seedu.address.model.entity.MatchableModule;
-import seedu.address.model.entity.Module;
+import seedu.address.model.Name;
+import seedu.address.model.deadline.Deadline;
+import seedu.address.model.event.Event;
+import seedu.address.model.event.EventType;
+import seedu.address.model.event.MatchableEvent;
+import seedu.address.model.module.MatchableModule;
+import seedu.address.model.module.Module;
+import seedu.address.model.module.ModuleCode;
 
 
 /**
@@ -39,9 +40,9 @@ public class AddDeadlineCommandParser implements Parser<AddDeadlineCommand> {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddDeadlineCommand.MESSAGE_USAGE));
         }
 
-        String moduleCode = ParserUtil.parseModuleCode(argMultimap.getValue(PREFIX_MODULE).get());
-        CalendarItemName name = ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
-        CalendarItemName eventName = ParserUtil.parseName(argMultimap.getValue(PREFIX_EVENT).get());
+        ModuleCode moduleCode = ParserUtil.parseModuleCode(argMultimap.getValue(PREFIX_MODULE).get());
+        Name name = ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
+        Name eventName = ParserUtil.parseName(argMultimap.getValue(PREFIX_EVENT).get());
         boolean isRepeated = ParserUtil.parseRepeat(argMultimap.getValue(PREFIX_REPEAT).get());
 
         // Todo: need to parse eventName to get eventType
